#130: Refactor product reference replacement utils unit test and provide unit tests for VariantRefReplacementUtils.
